Dr. Dobbs Header file for C99 and C++ source compt
From Dr.Dobb's
"Complex Arithmetic: In the Intersection of C and C++"
By Randy Meyers and Dr. Thomas Plum, March 01, 2003

http://www.drdobbs.com/complex-arithmetic-in-the-intersection-o/184401628

The listing does not carry a license, however the education spirit of
the writing seems to suggest that it is intended for use by readers:

	This month's installment does something a little different.
	Rather than just discuss a new feature in C99 and occasionally
	compare it to C++, we are going to concentrate on the
	intersection between C99 and C++. We are going to discuss how
	to write programs that use complex arithmetic and compile
	equally well under either C++ or C99. The key to opening the
	large intersection between C and C++ complex is the small
	header file named math-complex.h (shown in Listing 1).

This file is the contents of the "Listing 1" link.